refactor move TabConnectionErrorBanner.tsx
