+++ src/pageEditor/layout/PanelContent.tsx
@@ -22,7 +22,7 @@ import { tabStateActions } from "@/pageEditor/store/tabState/tabStateSlice";
 import { persistor } from "@/pageEditor/store/store";
 import { ModalProvider } from "@/components/ConfirmationModal";
 import ErrorBoundary from "@/components/ErrorBoundary";
-import TabConnectionErrorBanner from "@/pageEditor/TabConnectionErrorBanner";
+import TabConnectionErrorBanner from "@/pageEditor/components/TabConnectionErrorBanner";
 import RequireAuth from "@/auth/RequireAuth";
 import LoginCard from "@/pageEditor/components/LoginCard";
 import EditorLayout from "@/pageEditor/layout/EditorLayout";
